fix: mentor's comments
