+++ src/contexts/theme-context.tsx
@@ -2,8 +2,10 @@ import { createContext } from 'react';
 
 interface ThemeContextType {
   nightTheme: boolean;
+  toggleTheme: () => void;
 }
 
 export const ThemeContext = createContext<ThemeContextType>({
   nightTheme: false,
+  toggleTheme: () => {},
 });
