+++ src/shared/ui/selected-card-flyout/selected-card-flyout.scss
@@ -1,4 +1,4 @@
-@use '../../../../styles/scss/variables' as *;
+@use '../../../styles/scss/variables' as *;
 
 .hidden {
   visibility: hidden;
