+++ src/store/selectors/selectedCardSelector.ts
@@ -1,0 +1,3 @@
+import type { RootState } from '@/tests/test-utils/utils/setup-store';
+
+export const getSelectedCards = (state: RootState) => state.selectedCard.cards;
