+++ src/styles/style.scss
@@ -42,7 +42,6 @@ a:visited {
 
   width: 100%;
   height: 100%;
-  padding: 15px;
 
   font-family: 'Cera Pro', sans-serif;
 }
